fix: Corrige deleção em cascata

+++ api/app/routes/TraducaoRoute.py
@@ -502,6 +502,7 @@ def apagar_traducao(traducao_id):
 @verificar_professor_admin
 def remover_midia(traducao_id):
     body = request.get_json()
+    print(f"DEBUG: O que o Flask recebeu? {body}")
     if not body:
         return jsonify({"erro": "Corpo da requisição vazio"}), 400
 
@@ -518,31 +519,36 @@ def remover_midia(traducao_id):
         return jsonify({"erro": "Tradução não encontrada"}), 404
 
     #LOGICA DE EXCLUSÃO PERMANENTE
-    if apagar_servidor:
-        url_midia = doc.to_dict().get(tipo_midia)
-        if url_midia:
-            try:
-                bucket = storage.bucket()
-                parsed_url = urlparse(url_midia)
-                caminho_blob = unquote(parsed_url.path) 
-                prefixo_bucket = f"/{bucket.name}/"
-                if caminho_blob.startswith(prefixo_bucket):
-                    caminho_blob = caminho_blob.replace(prefixo_bucket, "", 1)
-                elif caminho_blob.startswith("/"):
-                    caminho_blob = caminho_blob.lstrip("/")
-                
-                #aponta para o arquivo no Storage
-                blob = bucket.blob(caminho_blob)
-                
-                if blob.exists():
-                    blob.delete()
-                    print(f"Arquivo apagado fisicamente do Storage: {caminho_blob}")
-                else:
-                    print("O arquivo não foi encontrado no Storage, mas será desvinculado do banco.")
+    url_midia = doc.to_dict().get(tipo_midia)
+    if apagar_servidor and url_midia:
+        try:
+            bucket = storage.bucket()
+            parsed_url = urlparse(url_midia)
+            caminho_blob = unquote(parsed_url.path).lstrip("/")
+            blob = bucket.blob(caminho_blob)
+            if blob.exists():
+                blob.delete()
+            
+            todas_traducoes = db.collection("traducao").stream()
+            
+            batch = db.batch()
+            contador = 0
+            
+            for doc_check in todas_traducoes:
+                dados = doc_check.to_dict()
+                valor_no_banco = dados.get(tipo_midia)
+                if valor_no_banco == url_midia:
+                    print(f"DEBUG: Encontrado match no documento: {doc_check.id}")
+                    ref = db.collection("traducao").document(doc_check.id)
+                    batch.update(ref, { tipo_midia: None })
+                    contador += 1
+            
+            if contador > 0:
+                batch.commit()
+            return jsonify({"mensagem": f"Mídia apagada e removida de {contador} tradução(ões)!"}), 200
                     
-            except Exception as e:
-                
-                return jsonify({"erro": f"Falha ao apagar arquivo do Storage: {str(e)}"}), 500
+        except Exception as e:
+            return jsonify({"erro": f"Falha ao apagar arquivo do Storage: {str(e)}"}), 500
     #LOGICA DE DESVINCULAR
     try:
         # Define o campo da mídia específica como nulo no banco de dados
